fix: address review feedback — regex bug, snippet dedup, imports
- Fix stateful regex bug in ArticleMarkdown: use odd-index check
  instead of regex.test() with g flag (coderabbit)
- Fix snippet boundary in articleUtils: use snippet.lastIndexOf(' ')
  without second arg since snippet is already sliced (coderabbit)
- Extract shared findSearchSnippet() utility in searchUtils to
  eliminate duplicated snippet logic between TruncatedText and
  truncateMarkdown
- Add word-boundary breaking to TruncatedText snippets
- Move React import to top of ArticleMarkdown
